results/rebuttal_apriori_unit_tests/run-20231114-021431_LLMatic-ZeroShot-CodeT-SelfRefine-Reflexion_donnemartin-system-design-oop-whatsapp_0_20-runs_log/whatsapp/3/LLMatic/message_service.py
class MessageService:
	def __init__(self):
		self.messages = {}
		self.read_receipts = {}
		self.encrypted_messages = {}
		self.images = {}
		self.contents = {}

	def send_message(self, sender_id, receiver_id, message):
		if receiver_id not in self.messages:
			self.messages[receiver_id] = []
		self.messages[receiver_id].append(message)
		return True

	def receive_message(self, receiver_id):
		if receiver_id in self.messages and self.messages[receiver_id]:
			return self.messages[receiver_id].pop(0)
		return None

	def mark_as_read(self, receiver_id, message_id):
		if receiver_id in self.read_receipts:
			self.read_receipts[receiver_id].append(message_id)
		return True

	def encrypt_message(self, sender_id, message):
		encrypted_message = f'Encrypted: {message}'
		self.encrypted_messages[sender_id] = encrypted_message
		return encrypted_message

	def decrypt_message(self, receiver_id, encrypted_message):
		if 'Encrypted: ' in encrypted_message:
			return encrypted_message.replace('Encrypted: ', '')
		return None

	def send_image(self, sender_id, receiver_id, image_path):
		if receiver_id not in self.images:
			self.images[receiver_id] = []
		self.images[receiver_id].append(image_path)
		return True

	def receive_image(self, receiver_id):
		if receiver_id in self.images and self.images[receiver_id]:
			return self.images[receiver_id].pop(0)
		return None

	def send_content(self, sender_id, receiver_id, content):
		if receiver_id not in self.contents:
			self.contents[receiver_id] = []
		self.contents[receiver_id].append(content)
		return True

	def receive_content(self, receiver_id):
		if receiver_id in self.contents and self.contents[receiver_id]:
			return self.contents[receiver_id].pop(0)
		return None
